pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agomtradepro"
version = "0.1.0"
description = "AgomTradePro investment decision platform"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "django>=5.2.11,<6.0",
    "djangorestframework",
    "django-filter",
    "django-environ",
    "django-cors-headers",
    "django-jazzmin>=3.0",
    "whitenoise>=6.8",
    "celery[redis,s3]",
    "redis",
    "django-celery-beat",
    "kombu",
    "openai>=1.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "statsmodels",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "tushare",
    "akshare==1.18.22",
    "feedparser",
    "drf-spectacular",
    "requests>=2.33.0",
    "urllib3>=2.6.3",
    "streamlit>=1.54.0",
    "plotly>=5.24",
    "prometheus-client>=0.19.0",
    "django-prometheus>=2.3.1",
    "gunicorn>=21.2",
    "python-dateutil>=2.8",
    "pytz>=2023.3",
    "typing-extensions>=4.9",
    "RestrictedPython",
    "sentry-sdk[django,celery]>=2.0",
    "cryptography>=46.0.6",
]

[project.optional-dependencies]
# Qlib AI 选股依赖（可选）
qlib = [
    "pyqlib>=0.9.0",
    "lightgbm>=4.0.0",
]

# Qlib 深度学习扩展（可选）
qlib-dl = [
    "pyqlib>=0.9.0",
    "lightgbm>=4.0.0",
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
]

# 开发工具
dev = [
    "pytest",
    "pytest-django",
    "pytest-cov",
    "pytest-mock",
    "pytest-asyncio>=0.21",
    "pytest-base-url>=2.1",
    "pytest-playwright>=0.5",
    "pytest-xdist>=3.5",
    "pytest-timeout>=2.3",
    "factory-boy>=3.3",
    "faker>=20.0",
    "playwright>=1.53",
    "black>=26.3.1",
    "isort",
    "ruff",
    "mypy",
    "pre-commit",
    "ipython",
    "ipdb",
]

# 完整开发环境（包含 Qlib）
all = [
    "pyqlib>=0.9.0",
    "lightgbm>=4.0.0",
    "pytest",
    "pytest-django",
    "pytest-cov",
    "pytest-mock",
    "pytest-asyncio>=0.21",
    "pytest-base-url>=2.1",
    "pytest-playwright>=0.5",
    "pytest-xdist>=3.5",
    "pytest-timeout>=2.3",
    "factory-boy>=3.3",
    "faker>=20.0",
    "playwright>=1.53",
    "black>=26.3.1",
    "isort",
    "ruff",
    "mypy",
    "pre-commit",
    "ipython",
    "ipdb",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["apps*", "core*", "shared*"]
exclude = [
    "tests*",
    "docs*",
    "sdk*",
    "deploy*",
    "docker*",
    "data*",
    "logs*",
    "reports*",
    "monitoring*",
    "streamlit_app*",
    "static*",
    "staticfiles*",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

# ── Existing debt: ignored for now, fix gradually ──────────────────────
# These rules trigger on existing code. New code must still pass them.
# Remove entries from this list as debt is paid off.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
# Legacy code with django.setup() before imports (tests only)
"tests/**" = ["E402"]
# Generated / third-party-adjacent code
"**/migrations/**" = ["E501", "F401", "F811"]
"docs/**" = ["F401", "F821", "E402"]
"sdk/**" = ["F401", "F821"]
# Scripts are standalone, not imported
"scripts/**" = ["F401", "F821", "E402"]
"tools/**" = ["F401", "F821", "E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "tushare.*",
    "akshare.*",
    "celery.*",
    "rest_framework.*",
    "drf_spectacular.*",
]
ignore_missing_imports = true
